Refactor sorting logic to hook
