+++ test/components/Sorteringsrad.test.tsx
@@ -7,8 +7,7 @@ describe('Sorteringsrad', () => {
   it('Skal rendre navn, fodselsnummer, virksomhet og veileder', () => {
     render(<Sorteringsrad onSortClick={() => void 0} />);
 
-    expect(screen.getByText('Etternavn')).to.exist;
-    expect(screen.getByText(', Fornavn')).to.exist;
+    expect(screen.getByText('Navn')).to.exist;
     expect(screen.getByText('Fødselsnummer')).to.exist;
     expect(screen.getByText('Virksomhet')).to.exist;
     expect(screen.getByText('Veileder')).to.exist;
